backend/requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
pytest>=7.0
pytest-asyncio>=0.21.0
requests>=2.28.0
beautifulsoup4>=4.11.0
cohere>=4.0.0
qdrant-client>=1.3.0
lxml>=4.9.0